% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige.R
\docType{package}
\name{krige}
\alias{krige}
\title{krige: Geospatial Kriging with Metropolis Sampling}
\description{
Estimates kriging models for geographical point-referenced data. Method is described '
  in Gill (2020) <doi:10.1177/1532440020930197>.
}
